+++ src/test/java/com/sonarorw/rules/S1197/S1197_compliant_official.java
@@ -1,0 +1,9 @@
+package com.sonarorw.rules.S1197;
+
+/**
+ * Compliant solution for S1197
+ */
+public class S1197_compliant_official {
+    // Compliant: Array designator is on the type.
+    int[][] matrix;
+}