+++ src/test/java/com/sonarorw/rules/S1659/S1659_compliant_official.java
@@ -1,0 +1,14 @@
+package com.sonarorw.rules.S1659;
+
+/**
+ * Compliant solution for S1659
+ */
+public class S1659_compliant_official {
+    // Compliant: Each variable is declared on its own line.
+    private int a;
+    private int b;
+    public void method() {
+        int c;
+        int d;
+    }
+}